pyproject.toml
[tool.poetry]
name = "dn42search"
version = "0.1.0"
description = ""
authors = ["Lukas Martini <hello@lutoma.org>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
Django = "^4.0.2"
django-q = "^1.3.9"
beautifulsoup4 = "^4.10.0"
psycopg2 = "^2.9.3"
robotspy = "^0.7.0"
requests-toolbelt = "^0.9.1"
boilerpy3 = "^1.0.5"
pysolr = "^3.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
